[refac] move SuccessMessage.java, ErrorMessage.java
