+++ src/main/java/org/hankki/hankkiserver/auth/filter/JwtAuthenticationFilter.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.hankki.hankkiserver.auth.UserAuthentication;
 import org.hankki.hankkiserver.auth.jwt.JwtProvider;
 import org.hankki.hankkiserver.auth.jwt.JwtValidator;
-import org.hankki.hankkiserver.common.dto.ErrorMessage;
+import org.hankki.hankkiserver.common.code.ErrorMessage;
 import org.hankki.hankkiserver.exception.UnauthorizedException;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
