+++ src/main/java/org/hankki/hankkiserver/common/code/ErrorMessage.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.common.dto;
+package org.hankki.hankkiserver.common.code;
 
 import lombok.AccessLevel;
 import lombok.Getter;
