+++ src/main/java/org/hankki/hankkiserver/exception/EntityNotFoundException.java
@@ -1,6 +1,6 @@
 package org.hankki.hankkiserver.exception;
 
-import org.hankki.hankkiserver.common.dto.ErrorMessage;
+import org.hankki.hankkiserver.common.code.ErrorMessage;
 
 public class EntityNotFoundException extends BusinessException {
     public EntityNotFoundException() {
